+++ src/main/java/com/momsme/momsme/address/service/VillageResponse.java
@@ -1,4 +1,4 @@
 package com.momsme.momsme.address.service;
 
-public record VillageResponse(Integer villageId, String villageName) {}
+public record VillageResponse(Long villageId, String villageName) {}
